refactor: move useSetupAdditionalProgressBarProps to proper place
